requirements.txt
APScheduler
email-validator
Flask
Flask-Limiter
Flask-Mail
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
geopy
gunicorn
Pillow
psycopg2-binary
pytest
python-dotenv
pytz
pyzbar
qrcode
reportlab
requests
SQLAlchemy
Werkzeug
WTForms
python-docx
